Initial commit ; both parts of the first 2 days are implemented.
This is one giant container for the Advent of Code 2020 challenge.
It is mostly a Kotlin learning project.

Hopefully by the end of the challenge the program itself will be able
to produce answers for any day, any part, using CLI arguments.
As it stands in this commit, code exists in the project for:

- Day 1 ; Part 2
- Day 2 : Part 1
- Day 2 : Part 2

Day 1, Part 1 will need to be partially re-implemented before
this project gets sealed.

I am also going to try to keep things well documented, so that
beginners at Kotlin (like me) can try to learn something from the
simple examples herein.