Implement Advent of Code day 3 - Slope Problem
Implements parts one and two for Day 3 of Advent
of Code 2020.

This adds a new class called Day3Grid, that can be used
to represent the grid in the problem. Supporting classes
are used to aid in the "wraparound" mechanic described.

An entrypoint called .simulate(,) is then used to calculate
the answers without having to re-instantiate the grid.